Refractor generate pdf
