+++ src/components/savedRecipesPage.tsx
@@ -3,6 +3,8 @@ import jsPDF from 'jspdf';
 import logo from '../img/cookingLabLogo2.png';
 import { SAVED_RECIPES_TITLE, SAVED_RECIPES_MSG, NO_SAVED_RECIPES, GENERATE_TXT, MODAL_TITLE, MODAL_TEXT } from '../i18n/constants';
 import CustomModal from './modal';
+import {handleGeneratePdfFile} from '../utils/index';
+
 
 const SavedRecipes = () => {
   const [savedRecipes, setSavedRecipes] = useState<Record<string, string>>(() => {
@@ -21,66 +23,15 @@ const SavedRecipes = () => {
   const handleShow = () => setShowModal(true);
   const handleClose = () => setShowModal(false);
 
-  const handleGeneratePdfFile = () => {
+  const generatePdf = () => {
     const recipes = Object.entries(savedRecipes);
     if (recipes.length === 0) {
       handleShow();
       setShowModal(true);
       return;
     }
-
-    const doc = new jsPDF();
-
-    doc.setFontSize(18);
-    doc.text('CookingLab Recipes', 105, 20, { align: 'center' });
-
-    const imgWidth = 50;
-    const imgHeight = 50;
-    doc.addImage(logo, 'PNG', 80, 30, imgWidth, imgHeight);
-
-    doc.setFontSize(14);
-    doc.text('Here are your saved recipes, click on the recipe name to view it:', 105, 90, { align: 'center' });
-
-    doc.setFontSize(12);
-    let yPosition = 100;
-    const pageWidth = doc.internal.pageSize.getWidth();
-    const margin = 10;
-    const maxWidth = pageWidth - margin * 2;
-
-    recipes.forEach(([name, url], index) => {
-      const text = `${index + 1}. ${name}`;
-
-      // Render the recipe name (split into multiple lines if necessary)
-      const splitText = doc.splitTextToSize(text, maxWidth);
-      splitText.forEach((line: string, lineIndex: number) => {
-        doc.textWithLink(line, margin, yPosition, { url });
-
-        // Only draw the underline starting after the number (index + 1) on the first line
-        if (lineIndex === 0) {
-          const numberWidth = doc.getTextWidth(`${index + 1}. `); // Calculate the width of the number
-          const textWidth = doc.getTextWidth(line) - numberWidth; // Calculate the width of the remaining text
-          doc.line(margin + numberWidth, yPosition + 1, margin + numberWidth + textWidth, yPosition + 1); // Draw underline only for the text
-        } else {
-          const textWidth = doc.getTextWidth(line); // For subsequent lines, underline the entire line
-          doc.line(margin, yPosition + 1, margin + textWidth, yPosition + 1);
-        }
-
-        yPosition += 10;
-
-        // Handle page overflow
-        if (yPosition > 280) {
-          doc.addPage();
-          yPosition = 20;
-        }
-      });
-    });
-
-    doc.setFontSize(10);
-    doc.text('Thank you for using CookingLab!', 105, 280, { align: 'center' });
-    doc.text('Visit us at: https://cooking-lab.netlify.app/', 105, 290, { align: 'center' });
-
-    doc.save('cooking_lab_recipes.pdf');
-  };
+    handleGeneratePdfFile(savedRecipes);
+  }
 
   return (
     <div className="App">
@@ -120,7 +71,7 @@ const SavedRecipes = () => {
                 <button
                   data-testid="save-recipe-btn"
                   className="btn btn-dark cooking-lab-btn me-3 ms-3 mt-3"
-                  onClick={() => handleGeneratePdfFile()}
+                  onClick={() => generatePdf()}
                 >
                   {GENERATE_TXT}
                 </button>
